+++ Process.h
@@ -77,7 +77,7 @@ public:
 	/** @return Total memory required by this process. */
 	size_t getMemoryRequired() const;
     size_t getAllocationBase() const;
-	void setAllocationBase(int base);
+	void setAllocationBase(size_t base);
     ///@/
 
 
@@ -178,6 +178,6 @@ private:
     /** @return Formatted timestamp. */
     std::string generateCreationTimestamp() const;
 
-	int allocationBase = -1;                                 // Base address for memory allocation
+	size_t allocationBase = size_t(-1);                                 // Base address for memory allocation
 	size_t memoryRequired;                                    // Total memory required by this process
 };
